Move Goobi to services.
